refactor: migrate hypav3 to svelte.ts and consolidate hypaV3 stores to hypaV3State
- Rename hypav3.ts to hypav3.svelte.ts
- Replace hypaV3ModalOpen, hypaV3ProgressStore with unified hypaV3State $state object
