+++ src/lib/Setting/Pages/OtherBotSettings.svelte
@@ -20,7 +20,7 @@
     import { getCharToken } from "src/ts/tokenizer";
     import { PlusIcon, PencilIcon, TrashIcon, DownloadIcon, HardDriveUploadIcon } from "@lucide/svelte";
     import { alertError, alertInput, alertConfirm, alertNormal } from "src/ts/alert";
-    import { createHypaV3Preset } from "src/ts/process/memory/hypav3";
+    import { createHypaV3Preset } from "src/ts/process/memory/hypav3.svelte";
 
     let submenu = $state(DBState.db.useLegacyGUI ? -1 : 0);
