+++ src/ts/stores.svelte.ts
@@ -66,13 +66,17 @@ export const alertStore = writable({
     type: 'none',
     msg: 'n',
 } as alertData)
-export const hypaV3ModalOpen = writable(false)
-export const hypaV3ProgressStore = writable({
+
+export const hypaV3State = $state({
     open: false,
-    miniMsg: '',
-    msg: '',
-    subMsg: '',
+    progress: {
+        open: false,
+        miniMsg: '',
+        msg: '',
+        subMsg: '',
+    }
 })
+
 export const selIdState = $state({
     selId: -1
 })
